feat: reconstuct user.controller
