+++ controllers/reply-controller.js
@@ -1,7 +1,5 @@
 const { Tweet, User, Reply } = require('../models')
 const Sequelize = require('sequelize')
-const { literal } = Sequelize
-const moment = require('moment')
 
 const replyController = {
   postComment: (req, res, next) => {
@@ -24,18 +22,20 @@ const replyController = {
   getComment: (req, res, next) => {
     const tweetId = req.params.tweet_id
     return Reply.findAll({
-    where: { tweetId },
-    order: [['createdAt', 'DESC']],
-    attributes: {
-      exclude: ['UserId', 'TweetId'],
-    },
-    raw: true
-  })
-  .then((replies) => {
-    if (!replies.length) throw new Error("Replies didn't exists!")
-    res.status(200).json(replies)
-  })
-  .catch(err => next(err))
+      include: [User],
+      where: { tweetId },
+      order: [['createdAt', 'DESC']],
+      attributes: {
+        exclude: ['UserId', 'TweetId'],
+      },
+      nest: true,
+      raw: true
+    })
+      .then((replies) => {
+        if (!replies.length) throw new Error("Replies didn't exists!")
+        res.status(200).json(replies)
+      })
+      .catch(err => next(err))
   },
   editComment: (req, res, next) => {
     const replyId = req.params.reply_id
@@ -44,59 +44,59 @@ const replyController = {
       raw: true,
       nest: true,
       attributes: {
-      exclude: ['UserId', 'TweetId'],
+        exclude: ['UserId', 'TweetId'],
       },
     })
-    .then(reply => {
-      if (!reply) throw new Error("Reply didn't exist!")
-      if (reply.userId !== userId) {
-        throw new Error('You are not authorized to edit this reply!')
-      }
-      return res.status(200).json(reply)
-    })
-    .catch(err => next(err))
+      .then(reply => {
+        if (!reply) throw new Error("Reply didn't exist!")
+        if (reply.userId !== userId) {
+          throw new Error('You are not authorized to edit this reply!')
+        }
+        return res.status(200).json(reply)
+      })
+      .catch(err => next(err))
   },
   putComment: (req, res, next) => {
     const replyId = req.params.reply_id
     const userId = req.user.id
     return Reply.findByPk(replyId)
-    .then(reply => {
-      if (!reply) {
-        throw new Error('Reply not found!')
-      }
-      if (reply.userId !== userId) {
-        throw new Error('You are not authorized to edit this reply!')
-      }
-      if (!req.body.comment) throw new Error('Comment text is required!')
-      
-      return reply.update({
-        comment: req.body.comment
+      .then(reply => {
+        if (!reply) {
+          throw new Error('Reply not found!')
+        }
+        if (reply.userId !== userId) {
+          throw new Error('You are not authorized to edit this reply!')
+        }
+        if (!req.body.comment) throw new Error('Comment text is required!')
+
+        return reply.update({
+          comment: req.body.comment
+        })
       })
-    })
-    .then(updatedReply => {
-      res.status(200).json(updatedReply)
-    })
-    .catch(err => next(err))
+      .then(updatedReply => {
+        res.status(200).json(updatedReply)
+      })
+      .catch(err => next(err))
   },
   deletedComment: (req, res, next) => {
     const replyId = req.params.reply_id
     const userId = req.user.id
     return Reply.findByPk(replyId)
-    .then(reply => {
-      if (!reply) {
-        throw new Error('Reply not found!')
-      }
-      if (reply.userId !== userId) {
-        throw new Error('You are not authorized to delete this reply!')
-      }
-      
-      return reply.destroy()
-    })
-    .then(deletedReply => {
-      res.status(200).json(deletedReply)
-    })
-    .catch(err => next(err))
+      .then(reply => {
+        if (!reply) {
+          throw new Error('Reply not found!')
+        }
+        if (reply.userId !== userId) {
+          throw new Error('You are not authorized to delete this reply!')
+        }
+
+        return reply.destroy()
+      })
+      .then(deletedReply => {
+        res.status(200).json(deletedReply)
+      })
+      .catch(err => next(err))
   }
-}  
+}
 
 module.exports = replyController
