+++ controllers/tweet-controller.js
@@ -1,6 +1,6 @@
 const { Tweet, User, Reply } = require('../models')
+const Sequelize = require('sequelize')
 const { literal } = Sequelize
-const moment = require('moment')
 
 const tweetController = {
   getTweets: (req, res, next) => {
@@ -10,17 +10,15 @@ const tweetController = {
       include: [{
         model: User,
         attributes: {
-        exclude: [ 'password', 'role'],
-        
+          exclude: ['password', 'role'],
         }
-      }, 
-    ],
-      attributes: { 
+      }],
+      attributes: {
         exclude: ['UserId'],
         include: [
           [literal('(SELECT COUNT(DISTINCT id) FROM Replies WHERE Replies.tweet_id = Tweet.id)'), 'replyCount'],
         ]
-     },
+      },
     })
       .then(tweet => {
         if (!tweet.length) throw new Error("Tweets didn't exist!")
@@ -28,7 +26,6 @@ const tweetController = {
       })
       .catch(err => next(err))
   },
-
   getTweet: (req, res, next) => {
     return Tweet.findByPk(req.params.tweet_id, {
       include: [Reply, User],
@@ -66,55 +63,55 @@ const tweetController = {
       raw: true,
       nest: true
     })
-    .then(tweet => {
-      if (!tweet) throw new Error("The tweet didn't exist!")
-      if (tweet.userId !== userId) {
-        throw new Error('You are not authorized to edit this tweet!')
-      }
-      return res.status(200).json(tweet)
-    })
-    .catch(err => next(err))
+      .then(tweet => {
+        if (!tweet) throw new Error("The tweet didn't exist!")
+        if (tweet.userId !== userId) {
+          throw new Error('You are not authorized to edit this tweet!')
+        }
+        return res.status(200).json(tweet)
+      })
+      .catch(err => next(err))
   },
   putTweet: (req, res, next) => {
     const tweetId = req.params.tweet_id
     const userId = req.user.id
     return Tweet.findByPk(tweetId)
-    .then(tweet => {
-      if (!tweet) {
-        throw new Error('Tweet not found!')
-      }
-      if (tweet.userId !== userId) {
-        throw new Error('You are not authorized to edit this tweet!')
-      }
-      if (!req.body.description) throw new Error('Description is required!')
-      
-      return tweet.update({
-        description: req.body.description
+      .then(tweet => {
+        if (!tweet) {
+          throw new Error('Tweet not found!')
+        }
+        if (tweet.userId !== userId) {
+          throw new Error('You are not authorized to edit this tweet!')
+        }
+        if (!req.body.description) throw new Error('Description is required!')
+
+        return tweet.update({
+          description: req.body.description
+        })
       })
-    })
-    .then(updatedTweet => {
-      res.status(200).json(updatedTweet)
-    })
-    .catch(err => next(err))
+      .then(updatedTweet => {
+        res.status(200).json(updatedTweet)
+      })
+      .catch(err => next(err))
   },
   deletedTweet: (req, res, next) => {
     const tweetId = req.params.tweet_id
     const userId = req.user.id
     return Tweet.findByPk(tweetId)
-    .then(tweet => {
-      if (!tweet) {
-        throw new Error('Tweet not found!')
-      }
-      if (tweet.userId !== userId) {
-        throw new Error('You are not authorized to delete this tweet!')
-      }
-      
-      return tweet.destroy()
-    })
-    .then(deletedTweet => {
-      res.status(200).json(deletedTweet)
-    })
-    .catch(err => next(err))
+      .then(tweet => {
+        if (!tweet) {
+          throw new Error('Tweet not found!')
+        }
+        if (tweet.userId !== userId) {
+          throw new Error('You are not authorized to delete this tweet!')
+        }
+
+        return tweet.destroy()
+      })
+      .then(deletedTweet => {
+        res.status(200).json(deletedTweet)
+      })
+      .catch(err => next(err))
   }
 }
